refactor(core): extract search algorithm to pure-Rust core-lib crate
Create rapid-fuzzy-core (crates/core-lib) as a standalone rlib with all
search algorithm logic: BigramIndex, PrecomputedSearch, search_impl, and
all supporting types. The napi crate is reduced to a thin binding layer
that wraps core-lib types with #[napi] annotations and From conversions.

This is the foundation for a future wasm-bindgen browser WASM target that
needs the same algorithms without napi-rs dependencies.

Closes #440
