+++ crates/core-lib/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "rapid-fuzzy-core"
+version = "1.1.1"
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+
+[lib]
+crate-type = ["rlib"]
+
+[dependencies]
+nucleo-matcher = { workspace = true }
